introduce Warden test helpers to feature specs
There is no need to always do a full auth cycle: visit new session path, fill in form and sign in.
In feature specs we focus on some particular feature, not the sign-in process.

So we should use `login_as` helper for all feature specs, except
those where we test (focus on) sign-in process.